1. build swift protobuf for distribution 2. ignore warnings in protobuf lib 3. enable -Wshorten-64-to-32 4. fixes some unused variable warnings

+++ src/Encrypt.cpp
@@ -89,7 +89,7 @@ Data AESCTREncrypt(const Data& key, const Data& data, Data& iv) {
     }
 
     Data result(data.size());
-    aes_ctr_encrypt(data.data(), result.data(), data.size(), iv.data(), aes_ctr_cbuf_inc, &ctx);
+    aes_ctr_encrypt(data.data(), result.data(), static_cast<int>(data.size()), iv.data(), aes_ctr_cbuf_inc, &ctx);
     return result;
 }
 
@@ -100,7 +100,7 @@ Data AESCTRDecrypt(const Data& key, const Data& data, Data& iv) {
     }
 
     Data result(data.size());
-    aes_ctr_decrypt(data.data(), result.data(), data.size(), iv.data(), aes_ctr_cbuf_inc, &ctx);
+    aes_ctr_decrypt(data.data(), result.data(), static_cast<int>(data.size()), iv.data(), aes_ctr_cbuf_inc, &ctx);
     return result;
 }
